requirements.txt
appdirs==1.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
certifi==2017.1.23
click==6.7
decorator==4.0.11
docutils==0.13.1
enum34==1.1.6
Flask==0.12
Flask-WTF==0.14.2
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.4
lockfile==0.12.2
luigi==2.5.0
MarkupSafe==0.23
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.2.0
pymongo==3.4.0
pyparsing==2.1.10
python-daemon==2.1.2
python-dateutil==2.6.0
scandir==1.4
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.15
WTForms==2.1
